updated dependency to inlucde spring web data jpa security

+++ src/main/java/com/project/ddm/repository/StationRepository.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StationRepository extends JpaRepository<Station, Long> {
-
+public interface StationRepository extends JpaRepository<Station, Integer> {
     List<Station> findAll();
 }
